jam_codec

Struct CountedInput

pub struct CountedInput<'a, I: Input> { /* private fields */ }

A wrapper for Input which tracks the number fo bytes that are successfully read.

If inner Input fails to read, the counter is not incremented.

It can count until u64::MAX - 1 accurately then saturate.

Implementations

impl<'a, I: Input> CountedInput<'a, I>

pub fn new(input: &'a mut I) -> Self

Create a new CountedInput with the given input.

pub fn count(&self) -> u64

Get the number of bytes successfully read.