jam_codec

Trait DecodeLimit

pub trait DecodeLimit: Sized {
    // Required methods
    fn decode_with_depth_limit<I: Input>(
        limit: u32,
        input: &mut I,
    ) -> Result<Self, Error>;
    fn decode_all_with_depth_limit(
        limit: u32,
        input: &mut &[u8],
    ) -> Result<Self, Error>;
}

Extension trait to Decode for decoding with a maximum recursion depth.

Required Methods

fn decode_with_depth_limit<I: Input>( limit: u32, input: &mut I, ) -> Result<Self, Error>

Decode Self with the given maximum recursion depth and advance input by the number of bytes consumed.

If limit is hit, an error is returned.

fn decode_all_with_depth_limit( limit: u32, input: &mut &[u8], ) -> Result<Self, Error>

Decode Self and consume all of the given input data.

If not all data is consumed or limit is hit, an error is returned.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.